+++ build.dependencies.gradle
@@ -1,4 +1,4 @@
-repositories {
+    repositories {
     mavenCentral()
     google() // Needed for androidx
 }
